requirements.txt
flask==2.0.1
tensorflow==2.8.0
numpy==1.21.2
Pillow==8.3.1 